requirements.txt
Flask==3.1.1
flask-cors==5.0.1
pandas==2.2.3
python-dotenv==1.1.0
requests==2.32.3
